Revert ogr_feature back to 1.9.2
I've tried to modify OGRFeature to indicate which field of record has
been changed. Then, this information can be used to generate an
efficient Update clause for SetFeature() function. However, it doesn't
work stable and generates a lot of unexpected memory problem. So, I
revert it back, and choose to use Update clause for almost all fields.